Correct incorrect spawn logic
In this commit we move the player's spawn state to onPlayerSpawn's event
handler. This is done to closely replicate what SA-MP does.

We also rename checkIfLeftVehicleByOtherMeans to
checkAndUpdatePlayerStates to update the ped's state data to
'PLAYER_STATE_ONFOOT' if their current state is either
'PLAYER_STATE_SPAWNED' or 'PLAYER_STATE_NONE'. This is done since SA-MP
stores whether the player is on foot or not if ped foot sync data is
sent over the network from the client to the server. This has been
documented in the code with a comment.